refactor(frontend): decompose CourseDetail into view-scoped components
CourseDetail.tsx shrank from 618 lines to 171 lines. The page now acts
as a thin orchestrator that loads data and delegates rendering to small
components in `frontend/src/pages/Course/detail/`:

- types.ts           — shared types, `formatDate`, `getCohortEnrollmentStatus`
- CourseDetailSkeleton.tsx — loading skeleton
- CohortSelectModal.tsx    — multi-cohort enrollment picker modal
- MaterialsModal.tsx       — downloadable materials modal
- UpcomingEvents.tsx       — upcoming deadlines/events strip
- EnrolledHeader.tsx       — compact header for enrolled view
- ModuleList.tsx           — module cards with lock/complete indicators
- NotEnrolledView.tsx      — pre-enrollment landing view
- EnrolledView.tsx         — enrolled user's main view
- index.ts                 — barrel re-exports

No functional/behavioral changes. tsc, eslint, vitest (85/85), and
`npm run build` all pass.

Made-with: Cursor
