+++ frontend/src/pages/Course/detail/NotEnrolledView.tsx
@@ -1,0 +1,177 @@
+import { useState } from "react"
+import { Link } from "react-router-dom"
+import { ArrowLeft, CalendarDays, Clock, Users } from "lucide-react"
+import { Button } from "@/components/ui/button"
+import { Card, CardContent } from "@/components/ui/card"
+import { Badge } from "@/components/ui/badge"
+import { toProxyImage } from "@/lib/images"
+import type { Course, Cohort } from "@/types"
+import { formatDate, isEnrollableCohort } from "./types"
+import { CohortSelectModal } from "./CohortSelectModal"
+
+interface Props {
+  course: Course
+  cohorts: Cohort[]
+  isOwner: boolean
+  isSignedIn: boolean
+  enrolling: boolean
+  onEnroll: (cohortId?: string) => Promise<void> | void
+}
+
+export function NotEnrolledView({
+  course,
+  cohorts,
+  isOwner,
+  isSignedIn,
+  enrolling,
+  onEnroll,
+}: Props) {
+  const [cohortSelectModal, setCohortSelectModal] = useState(false)
+  const [selectedCohortId, setSelectedCohortId] = useState<string | null>(null)
+
+  const activeCohort = cohorts.find((c) => c.status === "active")
+  const enrollableCohorts = cohorts.filter(isEnrollableCohort)
+  const canEnroll = enrollableCohorts.length > 0 || cohorts.length === 0
+
+  const handleEnrollClick = () => {
+    if (enrollableCohorts.length === 0) {
+      void onEnroll(undefined)
+      return
+    }
+    if (enrollableCohorts.length === 1) {
+      const first = enrollableCohorts[0]
+      if (first) void onEnroll(first.id)
+      return
+    }
+    const first = enrollableCohorts[0]
+    if (first) setSelectedCohortId(first.id)
+    setCohortSelectModal(true)
+  }
+
+  const confirmCohort = () => {
+    if (!selectedCohortId) return
+    setCohortSelectModal(false)
+    void onEnroll(selectedCohortId)
+  }
+
+  return (
+    <div className="container mx-auto px-4 py-6 max-w-3xl">
+      <Link to="/">
+        <Button variant="ghost" size="sm" className="mb-4 h-8 text-xs">
+          <ArrowLeft className="h-3.5 w-3.5 mr-1.5" />
+          All Courses
+        </Button>
+      </Link>
+
+      {course.image_url && (
+        <div className="mb-6 w-full aspect-[16/9] overflow-hidden rounded-md border bg-muted">
+          <img
+            src={toProxyImage(course.image_url)}
+            alt={course.title}
+            loading="lazy"
+            decoding="async"
+            className="h-full w-full object-cover"
+            onError={(e) => {
+              ;(e.currentTarget.parentElement as HTMLElement).style.display = "none"
+            }}
+          />
+        </div>
+      )}
+
+      <h1 className="font-serif text-3xl sm:text-4xl font-bold tracking-tight mb-3 text-wrap-safe">
+        {course.title}
+      </h1>
+
+      {course.description && (
+        <p className="text-muted-foreground leading-relaxed mb-6 whitespace-pre-line text-wrap-safe">
+          {course.description}
+        </p>
+      )}
+
+      {activeCohort && (
+        <Card className="mb-6">
+          <CardContent className="py-4">
+            <div className="flex items-center gap-2 mb-1">
+              <CalendarDays className="h-4 w-4 text-primary" />
+              <span className="font-medium">{activeCohort.name}</span>
+              <Badge variant={activeCohort.status === "active" ? "success" : "info"}>
+                {activeCohort.status}
+              </Badge>
+            </div>
+            <p className="text-sm text-muted-foreground">
+              {formatDate(activeCohort.start_date)} &mdash; {formatDate(activeCohort.end_date)}
+            </p>
+            {activeCohort.enrollment_start && activeCohort.enrollment_end && (
+              <p className="text-xs text-muted-foreground mt-1 flex items-center gap-1">
+                <Clock className="h-3 w-3" />
+                Enrollment: {formatDate(activeCohort.enrollment_start)} &mdash;{" "}
+                {formatDate(activeCohort.enrollment_end)}
+              </p>
+            )}
+            {activeCohort.max_students && (
+              <p className="text-xs text-muted-foreground mt-1">
+                {activeCohort.student_count}/{activeCohort.max_students} students enrolled
+              </p>
+            )}
+          </CardContent>
+        </Card>
+      )}
+
+      {!activeCohort &&
+        cohorts.length === 0 &&
+        (course.enrollment_start || course.enrollment_end) && (
+          <div className="flex flex-wrap items-center gap-2 text-sm mb-6">
+            <CalendarDays className="h-3.5 w-3.5 text-muted-foreground" />
+            {course.enrollment_start && course.enrollment_end && (
+              <span className="text-muted-foreground text-xs">
+                Enrollment: {formatDate(course.enrollment_start)} &mdash;{" "}
+                {formatDate(course.enrollment_end)}
+              </span>
+            )}
+          </div>
+        )}
+
+      <div>
+        {isOwner ? (
+          <Link to={`/teacher/courses/${course.id}`}>
+            <Button size="lg" variant="outline">
+              Manage Course
+            </Button>
+          </Link>
+        ) : isSignedIn ? (
+          <div>
+            <Button onClick={handleEnrollClick} disabled={enrolling || !canEnroll} size="lg">
+              <Users className="h-4 w-4 mr-2" />
+              {!canEnroll
+                ? "Enrollment not available"
+                : enrolling
+                  ? "Enrolling..."
+                  : "Enroll in Course"}
+            </Button>
+            {!canEnroll && (
+              <p className="text-sm text-muted-foreground mt-2">
+                {cohorts.length > 0
+                  ? "Enrollment window is closed for all available cohorts."
+                  : "No cohorts are currently available for this course."}
+              </p>
+            )}
+          </div>
+        ) : (
+          <Link to="/login">
+            <Button size="lg">Sign in to Enroll</Button>
+          </Link>
+        )}
+      </div>
+
+      <CohortSelectModal
+        open={cohortSelectModal}
+        onClose={() => setCohortSelectModal(false)}
+        cohorts={enrollableCohorts}
+        selectedCohortId={selectedCohortId}
+        onSelect={setSelectedCohortId}
+        onConfirm={confirmCohort}
+        enrolling={enrolling}
+      />
+    </div>
+  )
+}
