+++ frontend/src/pages/Course/detail/index.ts
@@ -1,0 +1,4 @@
+export { CourseDetailSkeleton } from "./CourseDetailSkeleton"
+export { NotEnrolledView } from "./NotEnrolledView"
+export { EnrolledView } from "./EnrolledView"
+export type { CourseMaterial } from "./types"
